move auth and cache errors to domain folder
