+++ src/domain/mod.rs
@@ -1,2 +1,3 @@
 pub mod auth_service;
+pub mod model;
 pub mod repositories;
