+++ src/domain/model/auth.rs
