+++ src/domain/model/cache_errors.rs
