+++ src/domain/model/mod.rs
@@ -1,0 +1,2 @@
+pub mod auth;
+pub mod cache_errors;
