+++ src/repositories/auth_repository.rs
@@ -1,5 +1,5 @@
+use crate::domain::model::auth::AuthorizationError;
 use crate::domain::repositories::auth_repository::AuthRepository;
-use crate::model::auth::AuthorizationError;
 use crate::model::login_user::{LoginUserError, LoginUserRequest};
 use crate::model::register_user::{RegisterUserError, RegisterUserRequest};
 use crate::model::user::{FilteredUser, User};
